requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
paho-mqtt==1.6.1
pyserial==3.5
python-decouple==3.8
requests==2.28.2
requests-futures==1.0.0
urllib3==1.26.15